Moved view classes and some format changes
