+++ usaspending_api/download/management/commands/generate_zip.py
@@ -10,26 +10,26 @@ from usaspending_api.download.lookups import JOB_STATUS_DICT
 from usaspending_api.download.models import DownloadJob
 from usaspending_api.download.filestreaming import csv_generation
 
-DEFAULT_VISIBILITY_TIMEOUT = 60*30
+DEFAULT_VISIBILITY_TIMEOUT = 60 * 30
 
-logger = logging.getLogger('console')
+logger = logging.getLogger("console")
 
 
 class Command(BaseCommand):
-
     def handle(self, *args, **options):
         """Run the application."""
         queue = sqs_queue(queue_name=settings.BULK_DOWNLOAD_SQS_QUEUE_NAME)
 
-        write_to_log(message='Starting SQS polling')
+        write_to_log(message="Starting SQS polling")
         while True:
             second_attempt = True
             try:
                 # Grabs one (or more) messages from the queue
-                messages = queue.receive_messages(WaitTimeSeconds=10, MessageAttributeNames=['All'],
-                                                  VisibilityTimeout=DEFAULT_VISIBILITY_TIMEOUT)
+                messages = queue.receive_messages(
+                    WaitTimeSeconds=10, MessageAttributeNames=["All"], VisibilityTimeout=DEFAULT_VISIBILITY_TIMEOUT
+                )
                 for message in messages:
-                    write_to_log(message='Message Received: {}'.format(message))
+                    write_to_log(message="Message Received: {}".format(message))
                     if message.body is not None:
                         # Retrieve and update the job
                         download_job = DownloadJob.objects.filter(download_job_id=int(message.body)).first()
@@ -47,7 +47,7 @@ class Command(BaseCommand):
 
                 if download_job:
                     download_job.error_message = str(e)
-                    download_job.job_status_id = JOB_STATUS_DICT['failed' if second_attempt else 'ready']
+                    download_job.job_status_id = JOB_STATUS_DICT["failed" if second_attempt else "ready"]
                     download_job.save()
             finally:
                 # Set visibility to 0 so that another attempt can be made to process in SQS immediately, instead of
