+++ usaspending_api/download/v2/download_column_historical_lookups.py
@@ -12,66 +12,6 @@ Code to generate these from spreadsheets:
 tail -n +3 'usaspending_api/data/DAIMS_IDD_Resorted+DRW+KB+GGv7/D2-Award (Financial Assistance)-Table 1.csv' >
 d2_columns.csv
 
-
-def find_column(col_name, model_classes):
-    for (model_class, prefix) in model_classes:
-        if hasattr(model_class, col_name):
-            return '{}{}'.format(prefix, col_name)
-    return None
-
-query_paths = {'transaction': {'d1': {}, 'd2': {}}, 'award': {'d1': {}, 'd2': {}}}
-human_names = {'transaction': {'d1': [], 'd2': []}, 'award': {'d1': [], 'd2': []}}
-
-models_award_d1 = ((Award, ''), (TransactionNormalized, 'latest_transaction__'),
-                    (TransactionFPDS, 'latest_transaction__contract_data__'))
-models_transaction_d1 = ((TransactionNormalized, ''), (TransactionFPDS, 'contract_data__'), (Award, 'award__'))
-models_award_d2 = ((Award, ''), (TransactionNormalized, 'latest_transaction__'),
-                    (TransactionFABS, 'latest_transaction__assistance_data__'))
-models_transaction_d2 = ((TransactionNormalized, ''), (TransactionFABS, 'assistance_data__'), (Award, 'award__'))
-
-def set_if_found(dl_name, path, model, file):
-    if path:
-        if dl_name in ('fain', 'piid', 'uri'):
-            dl_name = 'award_id_' + dl_name
-        query_paths[model][file][dl_name] = path
-        human_names[model][file].append(dl_name)
-    else:
-        print('Not found: {}: {}: {}'.format(model, file, dl_name))
-
-for row in d1:
-    if len(row['Download Name']) <= 1:
-        continue
-    if len(row['Database Tag']) <= 1:
-        continue
-    if row['Award Level'] == 'Y':
-        path = find_column(row['Database Tag'], models_award_d1)
-        set_if_found(row['Download Name'], path, 'award', 'd1')
-    if row['Transaction Level'] == 'Y':
-        path = find_column(row['Database Tag'], models_transaction_d1)
-        set_if_found(row['Download Name'], path, 'transaction', 'd1')
-
-# no database tags supplied for d2
-for row in d2:
-    if len(row['Download Name']) <= 1:
-        continue
-    if row['Award Level?'] == 'Y':
-        path = find_column(row['Download Name'], models_award_d2)
-        if not path:
-            # try what it was for D1
-            path = query_paths['award']['d1'].get(row['Download Name'])
-            if path:
-                col_name = path.split('__')[-1]
-                path = find_column(col_name, models_award_d2)
-        set_if_found(row['Download Name'], path, 'award', 'd2')
-    if row['Transaction Level?'] == 'Y':
-        path = find_column(row['Download Name'], models_transaction_d2)
-        if not path:
-            # try what it was for D1
-            path = query_paths['transaction']['d1'].get(row['Download Name'])
-            if path:
-                col_name = path.split('__')[-1]
-                path = find_column(col_name, models_transaction_d2)
-        set_if_found(row['Download Name'], path, 'transaction', 'd2')
 """
 
 query_paths = {
